resolving build errors
