+++ apps/mmessier/app/api/experiences/[experienceID]/helper.ts
@@ -1,0 +1,8 @@
+import { Experience } from '../../../../types';
+import { experiences } from '../helper';
+
+export const getExperience = async (
+  experienceID: string
+): Promise<Experience | null> => {
+  return experiences.find(({ id }) => id === experienceID) || null;
+};
