+++ apps/mmessier/app/projects/use-query-parameters-demo/page.tsx
@@ -1,0 +1,4 @@
+const UseQueryParametersDemo = () => {
+  return null;
+};
+export default UseQueryParametersDemo;
